@@ -29,45 +29,45 @@
   },
   "homepage": "https://github.com/allanfish/elementui#readme",
   "devDependencies": {
-    "babel-cli": "^6.18.0",
+    "babel-cli": "^6.22.2",
     "babel-jest": "^18.0.0",
     "babel-loader": "^6.2.10",
-    "babel-preset-es2015": "^6.16.0",
-    "babel-preset-react": "^6.16.0",
-    "babel-preset-stage-1": "^6.3.13",
-    "babel-standalone": "^6.21.1",
+    "babel-preset-es2015": "^6.22.0",
+    "babel-preset-react": "^6.22.0",
+    "babel-preset-stage-1": "^6.22.0",
+    "babel-standalone": "^6.22.1",
     "core-js": "^2.4.1",
     "css-loader": "^0.26.1",
-    "element-theme-default": "^1.1.2",
-    "eslint": "^3.12.2",
-    "eslint-config-airbnb": "^13.0.0",
+    "element-theme-default": "^1.1.6",
+    "eslint": "^3.15.0",
+    "eslint-config-airbnb": "^14.1.0",
     "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^2.2.3",
-    "eslint-plugin-react": "^6.8.0",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.9.0",
-    "html-webpack-plugin": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "^4.0.0",
+    "eslint-plugin-react": "^6.9.0",
+    "extract-text-webpack-plugin": "^2.0.0-rc.3",
+    "file-loader": "^0.10.0",
+    "html-webpack-plugin": "^2.28.0",
     "jest": "^18.1.0",
     "marked": "^0.3.6",
-    "node-sass": "^4.1.1",
+    "node-sass": "^4.5.0",
     "raw-loader": "^0.5.1",
     "react-hot-loader": "^3.0.0-beta.6",
-    "react-test-renderer": "^15.4.1",
-    "sass-loader": "^4.1.1",
+    "react-test-renderer": "^15.4.2",
+    "sass-loader": "^5.0.0",
     "style-loader": "^0.13.1",
     "url-loader": "^0.5.7",
-    "webpack": "^1.14.0",
-    "webpack-dev-server": "^1.16.2"
+    "webpack": "^2.2.1",
+    "webpack-dev-server": "^2.3.0"
   },
   "dependencies": {
     "async-validator": "^1.6.8",
     "classnames": "^2.2.5",
-    "popper.js": "^0.6.4",
-    "react": "^15.4.1",
-    "react-addons-css-transition-group": "^15.4.1",
-    "react-addons-transition-group": "^15.4.1",
+    "popper.js": "^1.0.6",
+    "react": "^15.4.2",
+    "react-addons-css-transition-group": "^15.4.2",
+    "react-addons-transition-group": "^15.4.2",
     "react-click-outside": "^2.2.0",
-    "react-dom": "^15.4.1"
+    "react-dom": "^15.4.2"
   },
   "jest": {
     "testRegex": "/jest/.+\\.(js|jsx)$",
