@@ -3,6 +3,7 @@ import React from 'react';
 import { PropTypes, Component } from '../../../libs';
 import { debounce } from '../../../libs/utils';
 import { getRangeHours } from '../utils'
+import { Scrollbar } from '../../scrollbar';
 
 function withIndex(arr) {
   return arr.map((e, i) => (e, i))
@@ -33,6 +34,7 @@ function propsToState(props) {
   return state
 }
 
+const SCROLL_AJUST_VALUE = 85;
 export default class TimeSpinner extends Component {
   static get propTypes() {
     return {
@@ -98,7 +100,7 @@ export default class TimeSpinner extends Component {
   }
 
   _handleScroll(type) {
-    const value = Math.min(Math.floor((this.refs[type].scrollTop - 80) / 32 + 3), 59);
+    const value = Math.min(Math.floor((this.refs[type].refs.wrap.scrollTop - SCROLL_AJUST_VALUE) / 32 + 3), 59);
     this.handleChange(type, value)
   }
 
@@ -116,13 +118,13 @@ export default class TimeSpinner extends Component {
 
   _ajustScrollTop({hours, minutes, seconds}) {
     if (hours != null) {
-      this.refs.hours.scrollTop = Math.max(0, (hours - 2.5) * 32 + 80);
+      this.refs.hours.refs.wrap.scrollTop = Math.max(0, (hours - 2.5) * 32 + SCROLL_AJUST_VALUE);
     }
     if (minutes != null) {
-      this.refs.minutes.scrollTop = Math.max(0, (minutes - 2.5) * 32 + 80);
+      this.refs.minutes.refs.wrap.scrollTop = Math.max(0, (minutes - 2.5) * 32 + SCROLL_AJUST_VALUE);
     }
     if (seconds != null) {
-      this.refs.seconds.scrollTop = Math.max(0, (seconds - 2.5) * 32 + 80);
+      this.refs.seconds.refs.wrap.scrollTop = Math.max(0, (seconds - 2.5) * 32 + SCROLL_AJUST_VALUE);
     }
   }
 
@@ -131,57 +133,67 @@ export default class TimeSpinner extends Component {
     const {isShowSeconds} = this.props
 
     return (
-      <div className={this.classNames('el-time-spinner', {'has-seconds': isShowSeconds})}>
-        <div
+      <div className={this.classNames('el-time-spinner', { 'has-seconds': isShowSeconds })}>
+        <Scrollbar
           onMouseEnter={() => this.emitSelectRange('hours')}
-          onWheel={() => this.handleScroll('hours')}
+          onWheel={() => {
+            this.handleScroll('hours')
+          } }
+          ref="hours"
           className="el-time-spinner__wrapper"
-          ref="hours">
-          <ul className="el-time-spinner__list">
-            {
-              hoursList.map((disabled, idx) => {
+          wrapStyle={{ maxHeight: 'inherit' }}
+          viewClass="el-time-spinner__list"
+          viewComponent="ul"
+          >
+          {
+            hoursList.map((disabled, idx) => {
 
-                return (
-                  <li
-                    key={idx}
-                    onClick={() => this.handleChange('hours', idx, disabled)}
-                    className={this.classNames('el-time-spinner__item', {
-                      'active': idx === hours,
-                      'disabled': disabled
-                    })}
-                    >{idx}</li>)
-              })
-            }
-          </ul>
-        </div>
-        <div
+              return (
+                <li
+                  key={idx}
+                  onClick={() => this.handleChange('hours', idx, disabled)}
+                  className={this.classNames('el-time-spinner__item', {
+                    'active': idx === hours,
+                    'disabled': disabled
+                  })}
+                  >{idx}</li>)
+            })
+          }
+        </Scrollbar>
+        <Scrollbar
           onMouseEnter={() => this.emitSelectRange('minutes')}
           onWheel={() => this.handleScroll('minutes')}
+          ref="minutes"
           className="el-time-spinner__wrapper"
-          ref="minutes">
-          <ul className="el-time-spinner__list">
-            {
-              minutesLisit.map((disabled, idx) => {
-                return (
-                  <li
-                    key={idx}
-                    onClick={() => this.handleChange('minutes', idx)}
-                    className={this.classNames('el-time-spinner__item', {
-                      'active': idx === minutes,
-                    })}
-                    >{idx}</li>
-                )
-              })
-            }
-          </ul>
-        </div>
+          wrapStyle={{ maxHeight: 'inherit' }}
+          viewClass="el-time-spinner__list"
+          viewComponent="ul"
+          >
+          {
+            minutesLisit.map((disabled, idx) => {
+              return (
+                <li
+                  key={idx}
+                  onClick={() => this.handleChange('minutes', idx)}
+                  className={this.classNames('el-time-spinner__item', {
+                    'active': idx === minutes,
+                  })}
+                  >{idx}</li>
+              )
+            })
+          }
+        </Scrollbar>
         {
           isShowSeconds && (
-            <div
+            <Scrollbar
               onMouseEnter={() => this.emitSelectRange('seconds')}
               onWheel={() => this.handleScroll('seconds')}
+              ref="seconds"
               className="el-time-spinner__wrapper"
-              ref="seconds">
+              wrapStyle={{ maxHeight: 'inherit' }}
+              viewClass="el-time-spinner__list"
+              viewComponent="ul"
+              >
               <ul className="el-time-spinner__list">
                 {
                   secondsList.map((disabled, idx) => {
@@ -197,7 +209,7 @@ export default class TimeSpinner extends Component {
                   })
                 }
               </ul>
-            </div>
+            </Scrollbar>
           )
         }
       </div>
