@@ -226,7 +226,10 @@ export default class DatePanel extends Component {
     const yearTranslation = Locale.t('el.datepicker.year');
     if (currentView === 'year') {
       const startYear = Math.floor(year / 10) * 10;
-      return startYear + ' ' + yearTranslation + '-' + (startYear + 9) + ' ' + yearTranslation;
+      if (yearTranslation){
+        return startYear + ' ' + yearTranslation + '-' + (startYear + 9) + ' ' + yearTranslation;
+      }
+      return startYear + ' - ' + (startYear + 9);
     }
     return year + ' ' + yearTranslation;
   }
@@ -275,7 +278,7 @@ export default class DatePanel extends Component {
   }
 
   render() {
-    const {showTime, shortcuts, sidebar} = this.props
+    const {showTime, shortcuts} = this.props
     const {currentView, date} = this.state
     const {month} = deconstructDate(date)
     const t = Locale.t
@@ -285,13 +288,11 @@ export default class DatePanel extends Component {
       <div
         ref="root"
         className={this.classNames('el-picker-panel el-date-picker', {
-          'has-sidebar': !!sidebar || shortcuts,
+          'has-sidebar': shortcuts,
           'has-time': showTime
         })}>
 
         <div className="el-picker-panel__body-wrapper">
-          {/* this slot is presented in vue element src code, but not used anywhere.*/}
-          {/*<slot name="sidebar" className="el-picker-panel__sidebar"></slot>*/}
           {
             Array.isArray(shortcuts) && (
               <div className="el-picker-panel__sidebar">
@@ -430,7 +431,6 @@ DatePanel.propTypes = {
   showTime: PropTypes.bool,
   showWeekNumber: PropTypes.bool,
   format: PropTypes.string,
-  // todo: constrait onClick method ? change to type: ()=>() ?
   // Array[{text: String, onClick: (picker)=>()}]
   shortcuts: PropTypes.arrayOf(
     React.PropTypes.shape({
@@ -438,7 +438,6 @@ DatePanel.propTypes = {
       // ()=>()
       onClick: PropTypes.func.isRequired
     })
-
   ),
   selectionMode: PropTypes.oneOf(Object.keys(SELECTION_MODES).map(e => SELECTION_MODES[e])),
   // (Date)=>bool, if true, disabled
