@@ -9,7 +9,7 @@ import Locale from '../../locale'
 const mapPropsToState = (props) => {
   const state = {
     format: props.format || 'HH:mm:ss',
-    currentDate: props.currentDate || Date()//todo: handle update.
+    currentDate: props.currentDate || Date()
   };
   state.isShowSeconds = (state.format || '').indexOf('ss') !== -1
   return state
